+++ src/main/java/com/timeToast/timeToast/dto/member/member/response/MemberAdminResponse.java
@@ -1,19 +1,22 @@
 package com.timeToast.timeToast.dto.member.member.response;
 
+import com.timeToast.timeToast.domain.enums.member.MemberRole;
 import com.timeToast.timeToast.domain.member.member.Member;
 import lombok.Builder;
 
 @Builder
-public record MemberManagerResponse(
+public record MemberAdminResponse(
         long memberId,
         String memberProfileUrl,
-        String nickname
+        String nickname,
+        MemberRole memberRole
 ) {
-    public static MemberManagerResponse from(final Member member) {
-        return MemberManagerResponse.builder()
+    public static MemberAdminResponse from(final Member member) {
+        return MemberAdminResponse.builder()
                 .memberId(member.getId())
                 .memberProfileUrl(member.getMemberProfileUrl())
                 .nickname(member.getNickname())
+                .memberRole(member.getMemberRole())
                 .build();
     }
 }
