+++ src/test/java/com/timeToast/timeToast/controller/event_toast/EventToastAdminControllerTest.java
@@ -1,8 +1,7 @@
 package com.timeToast.timeToast.controller.event_toast;
 
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
-import com.timeToast.timeToast.controller.eventToast.EventToastManagerController;
-import com.timeToast.timeToast.dto.event_toast.request.EventToastPostRequest;
+import com.timeToast.timeToast.controller.eventToast.EventToastAdminController;
 import com.timeToast.timeToast.dto.event_toast.request.EventToastRequest;
 import com.timeToast.timeToast.service.event_toast.EventToastService;
 import com.timeToast.timeToast.service.event_toast.EventToastServiceTest;
@@ -27,12 +26,12 @@ import static org.springframework.restdocs.request.RequestDocumentation.paramete
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public class EventToastManagerControllerTest extends BaseControllerTests {
+public class EventToastAdminControllerTest extends BaseControllerTests {
     private final EventToastService eventToastService = new EventToastServiceTest();
 
     @Override
     protected Object initController() {
-        return new EventToastManagerController(eventToastService);
+        return new EventToastAdminController(eventToastService);
     }
 
     @DisplayName("관리자는 이벤트 토스트의 목록을 조회할 수 있다.")
